Basic DMA remapping support via VT-d
This just sets up the VT-d DMAR units for the Linux cell. We do not yet
configure things for additional cells.

Also, although this feature will have to be mandatory for a real target,
we keep VT-d support optional in order to preserve the ability to run
Jailhouse inside QEMU/KVM.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>