driver/tools: Rework cell destruction API
Reference a cell to be destructed via its config file. The API to the
hypervisor is still based on names, but the driver IOCTL now takes a
reference to a fully loaded config.

This allows the driver to bring the CPUs that were used by the
destructed cell online again. It is also more handy for the user as
creation and destruction now both use the same token to reference a
cell.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>