x86: Park CPU in guest mode while waiting for SIPI
Instead of spinning inside the hypervisor while in wait-for-SIPI state,
introduce and use vmx_park_cpu. This functions stops the guest CPU in
halted state, interrupts disabled, so that only external events that
trigger a guest exit will wake it up again. This feature will also allow
us to keep the CPU for a longer period in the parked state without
burning lots of energy.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>