core: Generic memory region handling on cell destruction
Most of vmx_remap_to_linux and also the memory region loop in
vmx_cell_exit are generic and can be reused for adding VT-d support and,
later on, different architectures. So move the generic bits to the core
and provide arch_map_memory_region and arch_unmap_memory_region instead.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>