core: Allow a cell to declare itself terminated
This provides communication channel from the cell to the hypervisor to
signal cell termination, either regular one or after a failure. The cell
can set its cell_status field in the comm region for this purpose. The
hypervisor will take this field additionally into account when trying to
destroy a cell that requires managed shutdown.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>