core/configs: Rename cell flag "Unmanaged Exit" to "Passive Comm Region"
The purpose of this flag is to enable simple cells that do not actively
participate in the Communication Region protocols. The fact that we can
be shut down without prior consultation is one result of this but will
not be the only one in the future. Moreover, we may once use the Comm
Region to pass read-only information to the cells - without requiring
every cell to use the region actively or even map it.

Rename the flag to reflect its refined semantic.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>