core: Rename "Linux cell" to "root cell"
As we refer to the first cell that contains the boot-strap Linux as
"root cell" in the documentation, change the internal naming accordingly
in order to be more consistent. The affects linux_cell primarily, but
also a few related variable and function names as well as a couple of
console messages.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>