x86: Rework evaluation of MSR_IA32_VMX_TRUE_*_CTLS
The SDM recommends to keep default1-class controls enabled if they are
unknown to the VMM. This applies to most of those bits. Even worse, by
using the TRUE_*_CTLS, we kept DEBUG_CONTROLS saving/loading disabled on
most machines, corrupting the related states on vmexit.

Switch to the "untrue" capability MSRs, except for CR3 loading/storing,
which will ensure that default1 bits are kept enabled also on future
CPUs.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>