tools: config-create: Add support for using a data directory
Once installed, the jailhouse-config-create has to be able to use a data
directory for finding its templates. That dir will typically be
different from the one where the script is located. Prepare for this by
providing a datadir variable assignment that can be patched during
installation.

When running from the source tree, we continue to use that directory as
default template dir.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>