inmates: x86: Add E1000-based PCI DMA demonstration
This adds a simple Ethernet ping-ping via Intel E1000-based NICs. Run
two instances in the same network, and the first will become the test
controller while the second instance to show up will be the target.
Round trip times will be measured and dumped by the controller. This can
surely be further improved, and there are still some stability issues
(measurement stops on lost packet). It's a start.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>