core: Redefine PAGE_FLAG_UNCACHED to PAGE_FLAG_DEVICE
All (x86) users of this page flag map devices into the hypervisor
address space. We will do the same for ARM when mapping the debug UART.
For this we need a generic flag with the same semantics. As uncached is
different from device mappings, redefine the semantic of UNCACHED flag
for this purpose.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>