core/tools: Protect result of ARRAY_SIZE
Prevent any operator prioritization issues when embedding ARRAY_SIZE
into an expression by protecting the result with braces. This is a
defensive measure, no current user was affected by it.

Do not convert all the generated or copy&pasted ARRAY_SIZE defines in
the config files as long as they are not affected by the issue.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>