x86: vtd: Drop bogus checks in iommu_get_remapped_root_int
These checks make no sense: At this stage, we actually don't care if
there is a corresponding entry in the hypervisor remapping table for the
device. And the check of the vector number is actually comparing apples
(hypervisor-side base_index) to onions (guest-side irt_entries).
Finally, the tuple of device_id and vector stored in this function will
be validated by iommu_map_interrupt before use in case it gets forwarded
to it.

So drop this code. This will also make the function directly usable for
virtual devices that have no corresponding entry in the hypervisor's
interrupt remapping table.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>