Revert "core: Validate CPU ID before using it"
This reverts commit c1eb94e35d03a71bb528fe1a6ab5304e840f91b0.

This attempt to protect us from out-of-bounds CPU IDs is futile: already
arch_entry requires a usable per-cpu data structure to save and - in
case of an error - restore Linux state. It's the driver's duty to
prevent hypervisor invocations on out-of-bounds CPUs.

Moreover the patch contained a regression as it passed an uninitialized
member of cpu_data to cpu_id_valid.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>