x86: Introduce phys_ioapic abstraction
This structure will keep static and dynamic information about a physical
IOAPIC in a system. The three global variables ioapic_lock, ioapic and
shadow_redir_table are moved over, and an array of phys_ioapic
structures takes over their place. There is still only a single instance
supported, but once we have more, the physical base address will be used
to differentiate between them and also look them up from the array.

Internal functions of the IOAPIC subsystem are converted to make use of
the abstraction.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>