x86: Account for multiple IOAPICs per cell
Finally overcome the limitation of only one IOAPIC per cell, thus also
per system. We either look up the IOAPIC from the cell array based on
its physical address or we iterate over all IOAPICs of a cell when
needed - that's all. A good sign that we achieved this is the removal of
the IOAPIC_BASE_ADDR constant.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>