inmates: Power-up PHY on E1000 takeover
Clear the power-down bit in the PHY control register in case the
previous user turned it off. Linux does so since about 3.15.

Note that we do not try to reset the PHY. Getting it running again with
the proper link speed turned out to be too complicated (too many PHY
variants) for this little demo.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>