ci: Select single configuration for Coverity scan via branch name
As scan results of generic files are overwritten with the last
configuration build, allow to pick a specific config for stand-alone
analysis. This comes at the price of overwriting results on the project
page but is still better than missing something subtle.

To differentiate the snapshot in Coverity, patch the description that is
attached to the upload.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>